[rviz_interface] Move publishing methods to dedicated package.
