+++ architecture/package.xml
@@ -12,8 +12,8 @@
   <test_depend>gtest</test_depend>
   <depend>roscpp</depend>
   <depend>mavros_msgs</depend>
-  <depend>geometry_msgs</depend>
   <depend>std_msgs</depend>
+  <depend>geometry_msgs</depend>
   <depend>octomap_msgs</depend>
 
   <depend>architecture_msgs</depend>
@@ -22,6 +22,7 @@
   <depend>frontiers</depend>
   <depend>frontiers_msgs</depend>
   <depend>laser_free_to_cloud</depend>
+  <depend>rviz_interface</depend>
   <depend>tf</depend>
 
 </package>
