+++ frontiers/package.xml
@@ -37,6 +37,7 @@
     <depend>octomap_msgs</depend>
     <depend>octomap_ros</depend>
     <depend>visualization_msgs</depend>
+    <depend>rviz_interface</depend>
     <build_depend>catkin_simple</build_depend> 
   <!--   Note that this is equivalent to the following: -->
   <!--   <build_depend>roscpp</build_depend> -->
